콜로세움·로마 군단·노르드·베다 콘텐츠와 전투 규칙 정리
이번 세션 이전부터 작업 트리에 쌓여 있던 변경을 한데 담는다.
유닛·적·코드·스테이지 데이터와 그에 딸린 코드 구현, 원소 반응과
피해 태그 정리가 섞여 있다.

한 덩어리로 묶는 이유는 개별 주제로 가르려면 파일 하나 안에서 훅을
나눠야 하는데, 그렇게 쪼갠 커밋은 어느 것도 단독으로 빌드되지 않기
때문이다. 이 시점의 빌드는 통과한다(경고 0, 오류 0).

전투용 SD 스프라이트 경로는 없앴다. 해상도와 여백이 제각각이라 카드 안에서
크기가 들쭉날쭉했고, 연출 방향도 스프라이트가 아니라 카드가 반응하는 쪽으로
잡았다. 전용 임포터와 검증 스크립트도 함께 지운다.

+++ Assets/Scripts/Effects/Buffs/BuffEffects.cs
@@ -37,11 +37,12 @@ namespace Effects.Buffs
     public static class BuffStatus
     {
         /// <summary>
-        /// 단일 효과 버프 상태 생성. duration 0 이하 = 무한(라운드 종료까지).
+        /// 단일 효과 버프 상태 생성.
+        /// <paramref name="duration"/>은 <b>턴</b> 수이며 0 이하면 무한(라운드 종료까지)이다.
         /// </summary>
         public static UnitStatus Create(
             int id, string key, string name, Unit caster, Unit owner, BaseEffect effect,
-            float duration = -1f,
+            int duration = -1,
             BaseEnums.StatusStackPolicy stackPolicy = BaseEnums.StatusStackPolicy.Replace,
             BaseEnums.StatusCategory category = BaseEnums.StatusCategory.Positive,
             bool isBeneficial = false,
